enabale serde default when openapi type specifies default,
normally openapi would allow setting a default value, but implementing
this would be a huge hassle since serde expects the default value to be
passed as the result of a function. The capacity to add such a function
here and correctly refence it, would be a major challenge. Therefor we
just throw a warning and enable serde default for the field